Fly, Epser, Runner 클래스 삭제
Enum클래스에서 타입별로 move함수를 상속받아 재정의한다.
그러므로 Runner, Fly, Runner가 필요없어졌다, 삭제
